[#92205364] mv pivotal comment posting to lib
