+++ goship.go
@@ -12,7 +12,6 @@ import (
 	"log"
 	"math"
 	"net/http"
-	"net/url"
 	"os"
 	"os/exec"
 	"path"
@@ -622,7 +621,7 @@ func DeployHandler(w http.ResponseWriter, r *http.Request) {
 	}
 
 	if (c.Pivotal.Token != "") && (c.Pivotal.Project != "") && success {
-		err := postToPivotal(c.Pivotal, env, owner, name, toRevision, fromRevision)
+		err := goship.PostToPivotal(c.Pivotal, env, owner, name, toRevision, fromRevision)
 		if err != nil {
 			log.Println("ERROR: ", err)
 		} else {
@@ -682,72 +681,6 @@ func UnLockHandler(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 	http.Redirect(w, r, "/", http.StatusSeeOther)
-}
-
-func postToPivotal(piv *goship.PivotalConfiguration, env, owner, name, latest, current string) error {
-	gt := os.Getenv(gitHubAPITokenEnvVar)
-	t := &oauth.Transport{
-		Token: &oauth.Token{AccessToken: gt},
-	}
-	c := github.NewClient(t.Client())
-	comp, _, err := c.Repositories.CompareCommits(owner, name, current, latest)
-	if err != nil {
-		return err
-	}
-	pivRE, err := regexp.Compile("\\[.*#(\\d+)\\].*")
-	if err != nil {
-		return err
-	}
-	s := map[string]bool{}
-	layout := "2006-01-02 15:04:05"
-	timestamp := time.Now()
-	loc, err := time.LoadLocation("Asia/Tokyo")
-	if err != nil {
-		layout += " (UTC)"
-		log.Println("error: time zone information for Asia/Tokyo not found")
-	} else {
-		layout += " (JST)"
-		timestamp = timestamp.In(loc)
-	}
-	for _, commit := range comp.Commits {
-		cmi := *commit.Commit
-		cm := *cmi.Message
-		ids := pivRE.FindStringSubmatch(cm)
-		if ids != nil {
-			id := ids[1]
-			_, exists := s[id]
-			if !exists {
-				s[id] = true
-				m := fmt.Sprintf("Deployed to %s: %s", env, timestamp.Format(layout))
-				go PostPivotalComment(id, m, piv)
-			}
-		}
-	}
-	return nil
-}
-
-func PostPivotalComment(id string, m string, piv *goship.PivotalConfiguration) (err error) {
-	p := url.Values{}
-	p.Set("text", m)
-	req, err := http.NewRequest("POST", fmt.Sprintf(pivotalCommentURL, piv.Project, id), nil)
-	if err != nil {
-		log.Println("ERROR: could not form put request to Pivotal: ", err)
-		return err
-	}
-	req.URL.RawQuery = p.Encode()
-	req.Header.Add("Content-Type", "application/json")
-	req.Header.Add("X-TrackerToken", piv.Token)
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		log.Println("ERROR: could not make put request to Pivotal: ", err)
-		return err
-	}
-	defer resp.Body.Close()
-	body, _ := ioutil.ReadAll(resp.Body)
-	if resp.StatusCode != http.StatusOK {
-		log.Printf("ERROR: non-200 Response from Pivotal API: %s %s ", resp.Status, body)
-	}
-	return nil
 }
 
 func DeployPage(w http.ResponseWriter, r *http.Request) {
